+++ objs/tribal_ispc.h
@@ -21,10 +21,10 @@ namespace ispc { /* namespace */
 #if defined(__cplusplus) && !defined(__ISPC_NO_EXTERN_C)
 extern "C" {
 #endif // __cplusplus
-    extern void generate_triangles_and_velocities(double * lo, double * hi, int32_t nt, double * t[][3], double *  * v, uint32_t * tid, uint32_t * pid);
-    extern void generate_velocities(double * lo, double * hi, int32_t nt, double *  * v);
-    extern void integrate_triangles(double step, double * lo, double * hi, int32_t nt, double * t[][3], double *  * v);
-    extern void scale_triangles(double * lo, double * hi, int32_t nt, double * t[][3]);
+    extern void generate_triangles_and_velocities(double * lo, double * hi, int64_t nt, double * t[][3], double *  * v, uint64_t * tid, uint64_t * pid);
+    extern void generate_velocities(double * lo, double * hi, int64_t nt, double *  * v);
+    extern void integrate_triangles(double step, double * lo, double * hi, int64_t nt, double * t[][3], double *  * v);
+    extern void scale_triangles(double * lo, double * hi, int64_t nt, double * t[][3]);
 #if defined(__cplusplus) && !defined(__ISPC_NO_EXTERN_C)
 } /* end extern C */
 #endif // __cplusplus
